+++ doc/changes.rst
@@ -5,12 +5,16 @@ desitarget Change Log
 0.53.1 (unreleased)
 -------------------
 
-* No changes yet.
+* Don't insist that ``ZTILEID`` has to be in the ``zcat`` [`PR #687`_].
+* Install the SV2-related data files [`PR #686`_].
+
+.. _`PR #686`: https://github.com/desihub/desitarget/pull/686
+.. _`PR #687`: https://github.com/desihub/desitarget/pull/687
 
 0.53.0 (2021-03-18)
 -------------------
 
-* Update the ELG selection for SV2 [`PR #685`_]
+* Update the ELG selection for SV2 [`PR #685`_].
 * Implement full MTL loop [`PR #684`_]. Includes:
     * Modify ledgers based on any new tiles in a ``zcat`` directory.
     * An MTL tile file to track which tiles have been processed by MTL.